angle of arrows

+++ src/helpz/Constants.java
@@ -70,7 +70,7 @@ public class Constants {
             }
         }
 
-        public static float GetStartDmg(int towerType){
+        public static int GetStartDmg(int towerType){
             switch (towerType){
                 case CANNON:
                     return 25;
